pyproject.toml
[project]
name = "erpnext-dev-mcp"
version = "0.1.0"
description = "MCP server for ERPNext development"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=3.0.0",
    "httpx",
    "python-dotenv",
    "paramiko",
]

[project.scripts]
erpnext-mcp = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
